% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flfosr.R
\name{flfosr}
\alias{flfosr}
\title{Fast Longitudinal Function-On-Scalar Regression}
\usage{
flfosr(
  Y,
  X,
  z,
  k = 10,
  S = 2000,
  S_burn = S/2,
  varhyp = list(a_a = 0.1, b_a = 0.1, a_g = 0.1, b_g = 0.1, a_w = 0.1, b_w = 0.1)
)
}
\arguments{
\item{Y}{matrix of functional responses. Each column corresponds to one curve from a subject and each row is a measurement at a single time point on a common grid.}

\item{X}{design matrix of fixed effects using \code{model.matrix}. It should include column of 1s for the intercept.}

\item{z}{vector specifying group memberships of each curve, e.g. if there are 20 subjects with 5 repeated measurements each, \code{z=rep(1:20, each = 5)}.}

\item{k}{number of basis functions}

\item{S}{number of total MCMC iterations}

\item{S_burn}{number of initial MCMC iterations to discard for warm-up}

\item{varhyp}{a list containing hyperparameter values for the gamma variance priors. It must contain the following named elements:
\itemize{
\item a_a, b_a: shape and rate hyperparameters for fixed effects variance.
\item a_g, b_g: shape and rate hyperparameters for between-groups variance.
\item a_w, b_w: shape and rate hyperparameters for within-group variance.
}}
}
\value{
A list with the following elements:
\itemize{
\item X: the inputted design matrix
}
}
\description{
Performs MCMC estimation for Bayesian longitudinal function-on-scalar regression.
}
\examples{
simdata <- sim_lfosr_data(N = 20, Mi = 5, L = 3, Tn = 100)

m1 <- flfosr(simdata$Y, simdata$X, simdata$z)

}
