% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lfosr_data.R
\name{sim_lfosr_data}
\alias{sim_lfosr_data}
\title{Simulated Longitudinal Functional Data}
\usage{
sim_lfosr_data(
  N,
  Mi,
  L,
  Tn,
  K = 5,
  sig_noise = 0.1,
  sig_with = 1,
  sig_bet = 1,
  sig_alpha = 1,
  seed = NULL
)
}
\arguments{
\item{N}{number of subjects.}

\item{Mi}{integer or vector of length N specifying the number of replicates of each subject.}

\item{L}{number of non-zero scalar covariates.}

\item{Tn}{total number of time points.}

\item{K}{number of basis functions in final resulting basis (with K-2 reparameterized P-spline functions).}

\item{sig_noise}{observation level variance.}

\item{sig_with}{within-curve smooth variance.}

\item{sig_bet}{between-subjects smooth variance.}

\item{sig_alpha}{fixed effects variance.}

\item{seed}{(optional) sets the seed.}
}
\value{
list with the following elements:
\itemize{
\item Y: simulated longitudinal functional data
\item X: simulated design matrix
\item z: vector specifying group memberships of each curve
\item alphaf_true: true functional regression coefficients from which the data was generated from
}
}
\description{
Generate simulated longitudinal functional data and scalar covariates.
}
\examples{
simdata <- sim_lfosr_data(N = 20, Mi = 5, L = 3, Tn = 100)

}
